Split up pennylane_extensions module
- QFunc class is moved to its own module
- all other functions/classes are regrouped by area in a new
  `api_extensions` sub-module

Changes to existing files are restricted to import & path adjustments.

The DifferentiableCompileError has been adjusted as a top-level import
since it is a user-facing exception.
