+++ .pylintrc
@@ -22,7 +22,7 @@ ignore=test,llvm-project,mlir-hlo,Enzyme,.git,__pycache__,build,doc
 
 # List of decorators that change a function's signature. Will ignore certain errors
 # like 'no-value-for-parameter' on functions with these decorators.
-signature-mutators=catalyst.pennylane_extensions.for_loop
+signature-mutators=catalyst.api_extensions.control_flow.for_loop
 
 [MESSAGES CONTROL]
